wip(#1696): roborev r3 narrowing in progress — UNCERTIFIED, committed as work-loss insurance
Committed by the lead, not the implementer, while the lane is PARKED on disk
(box2 /data at 90%, 29G free; both gates held at >=40G). The implementer was
told to commit and had not, and 331 insertions / 331 deletions across 6 files
were sitting uncommitted where a watchdog kill or a full disk would lose them.

State: the R3-2/R3-3 narrowing (fraction-only validation at each public open
boundary; full Config::validate() removed from Database::open) is PARTIAL. This
commit has NOT been compiled or tested — no build has run since 5f1c3c0,
deliberately, because a build at this free-space level can hit ENOSPC and return
a red that reads as a code failure (box1: 24/31 components 'failed' on write
errors with tree-integrity PASS).

Do not treat this commit as certified anything. The next lite run after space is
restored is what certifies it.
